Update _analyze example to use POST
Appears to work with POST as well as GET and adding a body to a POST is more natural as a concept in general, but in particular for some clients.  As an example Postman limits adding a body to POST requests only (discussion here - https://github.com/postmanlabs/postman-app-support/issues/131).